% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.r
\name{colours_360}
\alias{colours_360}
\title{The 360info brand colours.}
\usage{
colours_360(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Strings of colours to request.
Includes helpers \code{blues}, \code{greens} and \code{greys}, which each give two
different colours. If no colours are specified, the entire palette is
returned.}
}
\value{
A named vector of colour values.
}
\description{
The 360info brand colours.
}
\examples{
colours_360("black", "blues")
colours_360(c("lightblue", "darkgrey"))
colours_360()
}
