% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{.onLoad}
\alias{.onLoad}
\title{On package startup, register the ttf fils in inst/extdata
as "Headline 360info", "Subhead 360info" and "Body 360info"}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{a character string giving the library directory where the package defining the namespace was found.}

\item{pkgname}{a character string giving the name of the package.}
}
\description{
(Note that the subhead font is no longer used by the theme
but is left for backwards-compatibility purposes.)
}
