% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.r
\name{extract_caption}
\alias{extract_caption}
\title{Take a ggplot2 object and return a list containing its title, subtitle, and
caption. Recognises Patchwork plots.
Based on https://github.com/grattan/grattantheme (under MIT licence)}
\usage{
extract_caption(p)
}
\arguments{
\item{p}{ggplot2 object}
}
\value{
The plot caption
}
\description{
Take a ggplot2 object and return a list containing its title, subtitle, and
caption. Recognises Patchwork plots.
Based on https://github.com/grattan/grattantheme (under MIT licence)
}
