% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.r
\name{fontfaces_360fonts}
\alias{fontfaces_360fonts}
\title{Returns a CSS \verb{@import} declaration for the 360 fonts. Used by \code{save_svg} so
that people viewing SVG plots on the web can see 360info fonts without
having them installed.}
\usage{
fontfaces_360fonts()
}
\description{
Returns a CSS \verb{@import} declaration for the 360 fonts. Used by \code{save_svg} so
that people viewing SVG plots on the web can see 360info fonts without
having them installed.
}
