% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.r
\name{get_360logo}
\alias{get_360logo}
\title{Get a sized and positioned rasterGrob of the 360info logo.}
\usage{
get_360logo(width, footer_prop_square)
}
\arguments{
\item{width}{The base physical plot width, as a \code{ggplot2::unit}}

\item{footer_prop}{The proportio nof the plot's height that will be taken
up the footer.}

\item{retina}{A multiplier for raster resolutions. Keeps relative text and
shape sizing intact.}
}
\value{
A sized and positioned \code{grid::rasterGrob} object displaying the logo
}
\description{
Get a sized and positioned rasterGrob of the 360info logo.
}
