% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.r
\docType{data}
\name{pal_360}
\alias{pal_360}
\title{The standard 360info colours.}
\format{
An object of class \code{character} of length 8.
}
\usage{
pal_360
}
\description{
The standard 360info colours.
}
\keyword{datasets}
