% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.r
\name{register_360fonts}
\alias{register_360fonts}
\title{Register specific Libre Franklin font weights for use in 360info plots. This
function runs on package load.}
\usage{
register_360fonts()
}
\description{
Register specific Libre Franklin font weights for use in 360info plots. This
function runs on package load.
}
