% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.r
\name{replace_caption}
\alias{replace_caption}
\title{Replace a ggplot2's labels (title, subtitle, and/or caption) with a
given string. Works for regular ggplot2 plots and patchwork plots.
Based on https://github.com/grattan/grattantheme (under MIT licence)}
\usage{
replace_caption(p, caption)
}
\arguments{
\item{p}{ggplot2 object}

\item{caption}{The caption to use in place of the plot's existing caption}
}
\description{
Replace a ggplot2's labels (title, subtitle, and/or caption) with a
given string. Works for regular ggplot2 plots and patchwork plots.
Based on https://github.com/grattan/grattantheme (under MIT licence)
}
