% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.r
\name{save_svg}
\alias{save_svg}
\title{Register specific Libre Franklin font weights for use in 360info plots. This
function runs on package load.}
\usage{
save_svg(plot, filename, height_factor = 1, scaling = 1.4, ...)
}
\arguments{
\item{plot}{The ggplot plot to save.}

\item{filename}{The path to the file to output.}

\item{height_factor}{How tall to render the graphic, relative to the width?
Common values are: 1 (square, the default), 9/16 (16:9 landscape), 16/9 (16:9 portrait), 3/4 (4:3 landscape), 4/3 (4:3 portrait)}

\item{scaling}{A scaling factor that affects the size of all text and lines.
By default, this is tuned to give the same font scaling as \code{save_png} (but
you may wish to increase it if you're making a particularly tall or short
graphic).}

\item{...}{Other arguments passed on to \code{svglite::svglite}.}
}
\value{
The original plot, invisibly (so you can use it in pipes)
}
\description{
Register specific Libre Franklin font weights for use in 360info plots. This
function runs on package load.
}
