% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.r
\name{textbox_360info}
\alias{textbox_360info}
\title{Add a 360info-themed textbox to the plot.}
\usage{
textbox_360info(
  box_style = c("light", "dark", "blue", "glass_light", "glass_dark"),
  ...
)
}
\arguments{
\item{box_style}{A string for one of the pre-defined 360info styles. Affects
the fill and text colour of the box. One of "light", "dark", "blue",
"glass_light" or "glass_dark".}

\item{...}{Other arguments passed on to \code{ggtext::geom_textbox} or
\code{ggplot2::geom_label}.}

\item{base_size}{Base font size, given in pts.}
}
\description{
Add a 360info-themed textbox to the plot.
}
