% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_360.r
\name{theme_360}
\alias{theme_360}
\title{Add 360info theming to a ggplot.}
\usage{
theme_360(base_family = "Body 360info", base_size = 14)
}
\arguments{
\item{base_family}{Name of the font family to use by default.}

\item{base_size}{Base font size, given in pts.}
}
\description{
Add 360info theming to a ggplot.
}
