% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_360info.r
\name{theme_360info}
\alias{theme_360info}
\title{Add 360info theming to a ggplot.}
\usage{
theme_360info(base_family = "Body 360info", base_size = 14)
}
\arguments{
\item{base_size}{Base font size, given in pts.}
}
\description{
Add 360info theming to a ggplot.
}
