% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.r
\name{get_360logo}
\alias{get_360logo}
\title{Get a sized and positioned rasterGrob of the 360info logo.}
\usage{
get_360logo(width, footer_width_prop)
}
\arguments{
\item{width}{The base physical plot width, as a \code{ggplot2::unit}}

\item{footer_width_prop}{The proportion of the plot's width that
will be taken up by the footer's height.}
}
\value{
A sized and positioned \code{grid::rasterGrob} object displaying the logo
}
\description{
Get a sized and positioned rasterGrob of the 360info logo.
}
