% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.r
\name{register_360fonts}
\alias{register_360fonts}
\title{Register specific Libre Franklin font weights for use in 360info plots. This
function runs on package load.}
\usage{
register_360fonts(
  pref = getOption("themes360info.franklin.pref", "itc"),
  reset_fonts = FALSE
)
}
\arguments{
\item{pref}{The preferred font to use: either \code{itc} for ITC Franklin Gothic,
or \code{libre} for Libre Franklin. If the preferred font is unavailable, the
other will be used as a fallback. By default, this argument is read from
the option "themes360info.franklin.pref", with \code{itc} used if it is unset.}

\item{reset_fonts}{If TRUE, reset the font register before registering
fonts. You may wish to use this if the wrong font was loaded, but it will
also clear any other custom fonts you've loaded with the \code{systemfonts}
package.}
}
\description{
Register specific Libre Franklin font weights for use in 360info plots. This
function runs on package load.
}
