% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.r
\name{save_360plot}
\alias{save_360plot}
\title{Save a graphic with the 360info logo in the lower right corner.}
\usage{
save_360plot(plot, filename, shape = "square", retina = 2, ...)
}
\arguments{
\item{plot}{The ggplot plot to save.}

\item{filename}{The path to the file to output.}

\item{shape}{What aspect ratio (width:height) should the graphic be?
Either a number specifying the ratio of the height to the width directly, or #' one of:
\itemize{
\item "phone-portrait", (9:16),
\item "photo-portrait" (2:3),
\item "sdtv-portrait" (3:4)
\item "square" (1:1, the default),
\item "sdtv-landscape" (4:3),
\item "photo-landcape" (3:2),
\item "phone-landscape" (16:9),
}}

\item{retina}{A scaling factor designed to give you a higher (or lower)
resolution graphic with the same relative size text and other elements.}

\item{...}{Other arguments (currently ignored).}
}
\value{
The original plot, invisibly (so you can use it in pipes)
}
\description{
Save a graphic with the 360info logo in the lower right corner.
}
